requirements.txt
Django==4.1.4
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-extensions==3.2.1
psycopg2-binary==2.9.1
phonenumbers==8.13.3
djangorestframework-simplejwt==5.2.2
django-environ==0.9.0
requests==2.28.2
redis==5.0.2
django-prometheus==2.2.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-django==0.44b0
opentelemetry-instrumentation-redis==0.44b0
opentelemetry-instrumentation-pymongo==0.44b0
opentelemetry-instrumentation-requests==0.44b0
opentelemetry-instrumentation-pika==0.44b0
opentelemetry-instrumentation-psycopg2==0.44b0
opentelemetry-instrumentation-mysqlclient==0.44b0
opentelemetry-exporter-otlp==1.23.0
pika==1.3.2
pymongo==4.6.2